Added self-hosted agent pool designation
This is because that a UI build will now go beyond 3 GB. Given that `git` will create an object pack file about the same size of the repo, the whole folder will be 2x of the size it should be. While removing `.git` could solve the building problem, `pm2` and version check still depend on the git commit hash. Therefore, simply removing `.git` directory won't work.